backend/requirements.txt
uvicorn==0.29.0
fastapi==0.110.1
pydantic==2.7.0
pytest==8.1.1
SQLAlchemy==2.0.29
PyMySQL==1.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
